pref: update build method
