+++ package.json
@@ -18,15 +18,12 @@
     "postinstall:electron": "node postinstall",
     "ng": "ng",
     "start": "npm run postinstall:electron && npm-run-all --parallel build:ng:debug electron:serve",
-    "build:electron:main": "tsc main.ts --outDir dist --target es6 --module commonjs && copyfiles package.json dist && cd dist && npm install --prod && cd ..",
+    "build:electron:main": "tsc scripts/main.ts --outDir dist --target es6 --module commonjs && tsc scripts/proxyServer.ts --outDir dist --target es6 --module commonjs",
     "build:ng:debug": "ng build --watch",
     "build:ng:prod": "ng build",
     "build": "npm run postinstall:electron && npm run build:ng:prod && npm run build:electron:main",
     "build:prod": "cross-env NODE_ENV=production npm run build",
-    "electron:serve": "npm run build:electron:main && electron ./dist --serve",
-    "electron:test": "electron ./dist",
-    "electron:dev": "npm run build && electron ./dist",
-    "electron:prod": "npm run build:prod && electron ./dist",
+    "electron:serve": "npm run build:electron:main && electron ./dist/main.js --serve",
     "electron:linux": "npm run build:prod && npx electron-builder build --linux",
     "electron:windows": "npm run build:prod && npx electron-builder build --windows",
     "electron:mac": "npm run build:prod && npx electron-builder build --mac",
